% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dr_estimator}
\alias{dr_estimator}
\title{dr_estimator: Calculate the doubly robust estimator proposed in Li, Goldberg, \& Zheng}
\usage{
dr_estimator(df, discretize_Z = F)
}
\arguments{
\item{df}{A dataframe with columns obseved time denoted by 'time', covariate 'Z'
'tau1' and 'tau2' which are the lower and upper times denoted by S and t in the paper,
and the cometing risk 'status' which is 0 for censoring, 1 for the event of interest, and 2 for the competing risk=status,}

\item{discretize_Z}{If 'TRUE', discretizes the covariate 'Z'. This can ease the calculation burden
for large dataframe 'df'. Default 'FALSE'. estimate the variance of the estimator. Default 'FALSE'}
}
\value{
A tibble with the estimate of the inverse-porbablity and the doubly roubst estimator for functions m_1 to m_6 discussed in the paper.
}
\description{
See details in the Method section
}
