% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\name{example1}
\alias{example1}
\title{example1: Create the data used in the first setting of the simulation section of Li, Goldberg, \& Zheng}
\usage{
example1(seed = 1, n = 500, censoring = "simple", to_round = T)
}
\arguments{
\item{seed}{The seed for the simulation}

\item{n}{The number of observations}

\item{censoring}{Either "simple" or "complex". Simple means that the censoring time is independent of the failure time. Complex means that it dependent.}

\item{to_round}{Boolean, If true, times are rounded down to two decimal places.}
}
\value{
A data framewith covariate 'Z',
observed time 'time'=Z,
the cometing risk 'status' which is 0 for censoring, 1 for the event of interest, and 2 for the competing risk=status,
the censoring time 'C', event of interest time 'T1', competing risk time 'T2',
and 'tau1' and 'tau2' which are the lower and upper times denoted by S and t in the paper
}
\description{
See details in the simulation section
}
