% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ipcw_estimator}
\alias{ipcw_estimator}
\title{ipcw_estimator: Calculate the ipcw estimator proposed in Li, Goldberg, \& Zheng}
\usage{
ipcw_estimator(df)
}
\arguments{
\item{df}{A dataframe with columns obseved time denoted by 'time', covariate 'Z'
'tau1' and 'tau2' which are the lower and upper times denoted by S and t in the paper,
and the cometing risk 'status' which is 0 for censoring, 1 for the event of interest, and 2 for the competing risk=status,}
}
\value{
A tibble with the estimate of the inverse-porbablity estimator for functions m_1 to m_6 discussed in the paper.
}
\description{
See details in the Method section
}
