% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary_statistics}
\alias{summary_statistics}
\title{summary_statistics: Calculate summary statistics TPP, FPP, FNP, TNP, Sensitivity, and Specificity}
\usage{
summary_statistics(est)
}
\arguments{
\item{est}{The result of either the ipcw_estimator or the dr_estimator}
}
\value{
A tibble with the estimate of the inverse-porbablity and the doubly roubst estimator for functions m_1 to m_6 discussed in the paper.
}
\description{
See details in the Method section
}
